[BUGFIX] LLXML Parser does not support localized label in the main LLXML
Many extension use the main LLXML file to provide default language and
others languages (fr, de, ...). The parser must support that.

This patch add this support. The priority is always given to the
language pack found in typo3conf/l10n. If a language pack is found
the localized label in the main LLXML file are not parsed.

Change-Id: I27e2ba03679ed4cf00c23c6140d243cf6aba2256
Resolves: #27972
Reviewed-on: http://review.typo3.org/3119
Reviewed-by: Björn Pedersen
Tested-by: Björn Pedersen
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers