[BUGFIX] Wrong check for existance of translated label
While tslib_fe::getLL() checks the existance of a translated label with
isset(), language::getLLL() does this with a strict string comparison.
This is evaluated as TRUE, if the subkey is not set at all, which is an
unexpected behavior.

Thus check, whether the label was translated into the current language
using isset().

Change-Id: I421d541d9aeaa2403052d212c7adba819a4e116f
Resolves: #28264
Releases: 4.6
Reviewed-on: http://review.typo3.org/3380
Reviewed-by: Xavier Perseguers
Reviewed-by: Stanislas Rolland
Tested-by: Stanislas Rolland
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert