[TASK] Replace module icon in language module
This patch replaces the icon in extensionmanager and admin tools menu
(module icon) with a more attractive one.
The new icon is from the http://www.famfamfam.com/lab/icons/silk/ silk
icon set, so it is free, credits to famfamfam are already give in
TYPO3 core, and it fits to the rest of the core icons.

Change-Id: Ic379c457b91eb5ba87d56d646a3b76a0909cb021
Resolves: #40546
Related: #39963
Releases: 6.0
Reviewed-on: http://review.typo3.org/14300
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn