[TASK] Remove "Clear language cache" menu item
* To show the item, SYS/lang/cache/clear_menu has to be manually
enabled, and a (not documented) userTSconfig option must be set.
* The item *never* worked since the underlying LanguageCacheClearer
class expects an array with cacheCmd set to either "all" or
"temp_cached", that is not set in the according AJAX call of the
menu action.
* Class LanguageCacheClearer hooks into DataHandler and is only
triggered on clear "all" cache.
* Clearing "all" cache flushes all caches anyway, so the hook
currently triggers flushing the language cache a second time.
* SYS/lang/cache/lifetime is unused and can be removed.

Since the functionality is broken, not documented, obsolete and
"internal" code of ext:lang, the whole thing is removed without
further deprecation.

Change-Id: Ic0d19bf31109832c9ed88d81d883516c109cf47e
Resolves: #52312
Releases: 6.2
Reviewed-on: https://review.typo3.org/24077
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters