[TASK] Cache menu needs clear namings and permissions
With the introduction of the grouping mechanisms of the
caching framework and the sensitivity of the system
caches group, the re-ordering and re-naming of cache
menu entries is necessary to ensure that admins and
editors use the caching mechanisms right.

Therefore there are new labels as well as "Clear all caches"
below "Clear page content" as it has a higher impact on
the system. The option to clear the Configuration Cache
/ System Caches is now disabled by default, even for
admins. It is active when the TSconfig option
options.clearCache.system = 1 is set or if the system is in
development context. This avoids slow page rendering
after clearing the system caches.

A new title attribute is added to explain the menu items
when hovering.

Resolves: #55119
Releases: 6.2
Change-Id: Ifd6026279a890b743c6c3cda09def19f9b7d59b7
Reviewed-on: https://review.typo3.org/26912
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Marcin S?gol
Tested-by: Marcin S?gol
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters