[BUGFIX] File upload directly in Doc-module is missing
Since the introduction of FAL the option to upload a file
directly in your form has disappeared.

With this change the upload posibility is back with some extras.
* progress is shown
* drag&drop of files
* multiple file upload

The options is can be disabled just like before in the
user settings and for each field in TCA.

foreign_selector_fieldTcaOverride.config.appearance.fileUploadAllowed = 0

The upload button is only shown for browsers that support
drag&drop upload. And only for FAL fields (sys_file_reference).

Resolves: #55545
Releases: 6.2
Change-Id: I4b984099095d7f66c3d37023cd6c547b2ff5d59f
Reviewed-on: https://review.typo3.org/27237
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack