[FEATURE] Usability improvements by top searchbox in list module
For a better user experience in the list module, it makes sense to display
the search box at the top and not at the bottom of the page.
If a user works with a lot of records in the list module it is really
annoying to scroll down, the position in the top is more ergonomic.

With the new search icon in the doc header it is possible to toggle the
new search toolbar. if the user enter a search term and the results in
the list module are filtered, the search toolbar is visible by default.

This patch works also with the searchbox in the page module. The
searchbox in the ElementBrowser is always visible, but also now on top
above the table.

Resolves: #59763
Documentation: #59763
Related: #61777
Releases: master
Change-Id: I25e81b5f5b914cb75bf30de16fcc2da9edd7f6e6
Reviewed-on: http://review.typo3.org/30997
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Felix Kopp <felix-source@phorax.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>