[FEATURE] Add autoplay option for filetype audio and video
With the RenderingRegistry introduced with #61800 we have the option
to render audio and video tags. #66366 adds a ViewHelper so you
can use the Renderer classes in FE.
What is missing is a convenient way for the editors to enable
autoplay for a audio or video file attached to a Content Element.

This change adds a autoplay checkbox to sys_file_reference and
enhances the AudioTagRenderer and VideoTagRenderer to use this
value when set.

Releases: master
Resolves: #66371
Change-Id: I4c551c0e966af406859641e21833eb12586c55e5
Reviewed-on: http://review.typo3.org/38625
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>