[BUGFIX] Prevent error for tree in flexform
This patch prevent JS errors in case a tree is rendered in a FlexForm,
before the record was saved for the first time. In this case the data
can’t be loaded and we display a message.

This is a bit hacky but at the moment the best way to fix the issue.

Resolves: #78350
Releases: master
Change-Id: Ic11ae6e5c22711fe9a10fe3287b015ea1d87a30f
Reviewed-on: https://review.typo3.org/50275
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>