[FEATURE] Show list of failed input fields in FormEngine
When validating input fields of the FormEngine fails, a button is now
rendered into the least possible button bar in the module document header.
Clicking the button renders a list of all input elements whose validation
failed.

Clicking onto a field in that list automatically focuses the field in the
form.

Resolves: #79521
Releases: master
Change-Id: I9e232f4d1b27216ccf4a1c7b88d4a9c70b49f4f0
Reviewed-on: https://review.typo3.org/51452
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>