[!!!][TASK] Always include pageId in cHash calculation
The option $TYPO3_CONF_VARS[FE][cHashIncludePageId]
was included as part of a security fix to stay backwards-compatible
with existing installations.

The option was only there as a intermediate step, however
the cHash calculation should always contain the pageID,
so the option is removed and should be enabled by default.

The change removes the option via a silent configuration
updater in the Install Tool, removes the status report and
changes the cHash calculation to always include a pageId,
and if the page ID is not given, an exception is now
thrown all the time.

Resolves: #80050
Releases: master
Change-Id: Iac8eef1273848309da62deb24160f1c14ef338f4
Reviewed-on: https://review.typo3.org/51883
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>