[BUGFIX] Add possibility to define crop variants per table
Currently it is not possible to define crop variants per
table or content type. This makes this feature a lot less useful
as it could be.

Add a form data provider to make this possible.

Also fix the issue that the full size image is used instead of
a smaller variant in the interface and hide the cropping
tool if no image size can be determined (happens for some SVGs).

Focus area now also works when no cover areas are defined.

Resolves: #80105
Resolves: #80157
Resolves: #79961
Releases: master
Change-Id: I3333a59c84b6f55f7bf300663ea2ff52bf70e698
Reviewed-on: https://review.typo3.org/51951
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>