[FEATURE] Add duplicate button
This patch provides a new "duplicate" button inside of the topbar of the
edit record form next to the close button. This button creates and opens
a new copy of the current element which is positioned right below the
existing one.
If there are unsaved changes in the currently open record, a modal
appears with options to cancel, dismiss the changes and clone or save
the changes and clone the changed record.

Resolves: #77685
Releases: master
Change-Id: I9e2a0f02fc9c4452f0a7d868ea17aeacb9375d95
Reviewed-on: https://review.typo3.org/53101
Reviewed-by: Wolfgang Klinger <wolfgang@wazum.com>
Tested-by: Wolfgang Klinger <wolfgang@wazum.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>