[BUGFIX] Don't download language packs of not loaded extensions
The backend language module struggles with downloading
language packs of existing, but not loaded extensions
and fetchess old (non core version specific) packs
for core extensions.
This is hard to solve on a bugfix level for v7 and v8,
and in general it does not make much sense to have
language packs of not loaded extensions lying around
in typo3conf/l10n.
The patch ignores fetching of language packs for
not loaded extensions, it easily applies to all
maintained core versions.

Change-Id: I9ad885012a572368f7946f1027d870ee09550034
Resolves: #83406
Releases: master, 8.7, 7.6
Reviewed-on: https://review.typo3.org/55863
Reviewed-by: Michael Stucki <michael.stucki@typo3.org>
Tested-by: Michael Stucki <michael.stucki@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>